requirements.txt
streamlit
google-auth-oauthlib
google-auth-httplib2
youtube-data-api